Move to pkg/naming from address.
