+++ pkg/naming/naming_test.go
@@ -1,4 +1,4 @@
-package address
+package naming
 
 import (
 	"testing"
